extensions/connectors/sql/pyproject.toml
[tool.poetry]
name = "pandasai-sql"
version = "0.1.1"
description = "SQL integration for PandasAI"
authors = ["Gabriele Venturi"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pandasai = ">=3.0.0a0"
sqlalchemy = "^2.0.0"
numpy = ">=1.23.2,<2.0.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"


[tool.poetry.group.dev.dependencies]
pymysql = "^1.1.1"
psycopg2-binary = "^2.9.10"
cockroachdb = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
